add dash template and update hatch and test-hatch
